+++ cfxplorer/version.py
@@ -1,6 +1,6 @@
 """
-``cfxplorer`` is a python package for
-generating counterfactual explanations for a tree-based model
+``CFXplorer`` is a python package for
+generating counterfactual explanations for a given model and feature set.
 """
 # PEP0440 compatible formatted version, see:
 # https://www.python.org/dev/peps/pep-0440/
@@ -18,4 +18,4 @@ generating counterfactual explanations for a tree-based model
 # Dev branch marker is: 'X.Y.dev' or 'X.Y.devN' where N is an integer.
 # 'X.Y.dev0' is the canonical version of 'X.Y.dev'
 #
-__version__ = "0.0.dev4"  # pragma: no cover
+__version__ = "0.0.dev7"  # pragma: no cover
